refactor: 랜덤 클래스를 LadderMaker 클래스의 메소드로 변경
