+++ src/main/java/ladder/model/LadderMaker.java
@@ -3,7 +3,9 @@ package ladder.model;
 import static ladder.util.validator.LadderValidator.checkEmpty;
 import static ladder.util.validator.LadderValidator.checkLadderNumberStandard;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Random;
 
 public class LadderMaker {
 
@@ -21,8 +23,20 @@ public class LadderMaker {
     }
 
     public List<Boolean> makeLadder(int personCount) {
-        RandomBoolean randomBoolean = new RandomBoolean();
-        Line lines = new Line(randomBoolean.createRandomColumn(personCount));
+        Line lines = new Line(createRandomColumn(personCount));
         return lines.getPoints();
+    }
+
+    public List<Boolean> createRandomColumn(int personCount) {
+        List<Boolean> columns = new ArrayList<>();
+        for (int i = 0; i < personCount - 1; i++) {
+            columns.add(getRandomBoolean());
+        }
+        return columns;
+    }
+
+    private boolean getRandomBoolean() {
+        Random random = new Random();
+        return random.nextBoolean();
     }
 }
